Fix compilation warning with x86_64-w64-mingw32 (use same macro definition as in libtiff tif_fax3.c)
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@26259 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/aigrid/aigccitt.c
@@ -1487,7 +1487,7 @@ typedef struct {
 
 #define	is2DEncoding(sp) \
 	(sp->b.groupoptions & GROUP3OPT_2DENCODING)
-#define	isAligned(p,t)	((((unsigned long)(p)) & (sizeof (t)-1)) == 0)
+#define	isAligned(p,t)	((((size_t)(p)) & (sizeof (t)-1)) == 0)
 
 /*
  * Group 3 and Group 4 Decoding.
